Rhythm Changes
Allowed a much larger span of time (or “swing”) of the player to hit a
rhythm. I had put systems in place before that I hadn’t even realized
to allow for this, now the player has slightly longer than a 32nd
before or after each marked timestamp to hit the rhythm.

also added details to the console view so you can have exact feed back
on your rhythm for both hit and missed rhythms. the console view will
be in this form:

Mintime (bottom of allotted range):
your time
Maxtime (top of alotted range):

known bugs - for some reason on rhythm sets with 7 indices, it is
impossible to attain a 100% pitch score even when all pitches are
perfect. I am currently looking into this.